x/dstorage/client/cli/query_stored_file.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/dfsapp/dstorage/x/dstorage/types"
	"github.com/spf13/cobra"
)

func CmdListStoredFile() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-stored-file",
		Short: "list all storedFile",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllStoredFileRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.StoredFileAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowStoredFile() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-stored-file [index]",
		Short: "shows a storedFile",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryGetStoredFileRequest{
				Index: argIndex,
			}

			res, err := queryClient.StoredFile(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
